refactor(project): Put voiture tampon in a sub folder
